trival: add Instaced key word for sub object
Subobject will need it to be saved correctly on disk, and we have
see that subobject on map lost subobjects after meta data change
in cpp file, so we add it for every subobject.

Also added 'Super' for constructors.

also moved out mainthread loop functions from runnable
to visual obj.

Signed-off-by: Wenchao Xia <wenchaoopensource@163.com>