core: adjust data replication function
The code that handle data replication and notification to visual layer,
is reorgnized to one function in core actor, which will handle
standalone, listenserver, dedicated server cases.

also renamed the view role to data role in player controller base, since
it represent the data's state, not the visual's state.

also fix a bug that visual client should not call coreDataPullLoopInner().

this patch may break build since it is committed in the middle of work.

Signed-off-by: Wenchao Xia <wenchaoopensource@163.com>