UI: map in world locations to UI borders
Now attender info widget is not static but dynamic mapped to UI
according to the their projected points.

To archieve so, interfaces are reorganized to keep cached data in
C++, and help retrive modelinfo in widget, and to keep widget
layout data comply with model info.

Bascically, the widget's modelinfo is directly calculated from
canvas slot layout data, and it's transform pivot and slot layout
alignment is same, and it is similar as origin point in 3D actor.
The widget itself just need to handle contentSize, not directly
handle modelinfo data.

One thing in mind, the parent need a chance to call API to make
widget layout data comform with model info data, right after
widget can tell it's content size. Since now there is no hook
on child widget when child is added to canvals pannel, we
need to do so for child widget, in parent umg class's init function
for static ones, and in parent create child widget function for
dynamic ones.

Signed-off-by: Wenchao Xia <wenchaoopensource@163.com>